requirements.txt
Flask
openai
whisper
line-bot-sdk
openpyxl
pandas
selenium
undetected-chromedriver
webdriver-manager